move util/* to the appropriate generator
