+++ generators/base-application/support/index.mts
@@ -17,7 +17,13 @@
  * limitations under the License.
  */
 
+export * from './debug.mjs';
 export * from './doc.mjs';
 export * from './enum.mjs';
 export * from './entities.mjs';
 export * from './field-utils.mjs';
+export { default as prepareEntity } from './prepare-entity.mjs';
+export * from './prepare-entity.mjs';
+export { default as prepareField } from './prepare-field.mjs';
+export * from './prepare-field.mjs';
+export { default as prepareRelationship } from './prepare-relationship.mjs';
