+++ generators/base-application/support/prepare-field.mjs
@@ -16,21 +16,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import assert from 'assert';
 import _ from 'lodash';
-import { databaseTypes, entityOptions, fieldTypes, reservedKeywords, validations } from '../jdl/jhipster/index.mjs';
-import { getTypescriptType } from '../generators/client/support/index.mjs';
-import { fieldIsEnum } from '../generators/base-application/support/index.mjs';
+import { fieldTypes, validations } from '../../../jdl/jhipster/index.mjs';
+import { getTypescriptType, prepareField as prepareClientFieldForTemplates } from '../../client/support/index.mjs';
+import { prepareField as prepareServerFieldForTemplates } from '../../server/support/index.mjs';
+import { fieldIsEnum } from './field-utils.mjs';
 
-const { isReservedTableName } = reservedKeywords;
 const { BlobTypes, CommonDBTypes, RelationalOnlyDBTypes } = fieldTypes;
 const {
   Validations: { MIN, MINLENGTH, MINBYTES, MAX, MAXBYTES, MAXLENGTH, PATTERN, REQUIRED, UNIQUE },
 } = validations;
-const { MYSQL, SQL } = databaseTypes;
-const { MapperTypes } = entityOptions;
 
-const { MAPSTRUCT } = MapperTypes;
 const { TEXT, IMAGE, ANY } = BlobTypes;
 const {
   BOOLEAN,
@@ -252,7 +248,7 @@ function _derivedProperties(field) {
   });
 }
 
-export function prepareFieldForTemplates(entityWithConfig, field, generator) {
+export default function prepareField(entityWithConfig, field, generator) {
   prepareCommonFieldForTemplates(entityWithConfig, field, generator);
 
   if (entityWithConfig.prodDatabaseType || entityWithConfig.databaseType) {
@@ -263,7 +259,7 @@ export function prepareFieldForTemplates(entityWithConfig, field, generator) {
   return field;
 }
 
-export function prepareCommonFieldForTemplates(entityWithConfig, field, generator) {
+function prepareCommonFieldForTemplates(entityWithConfig, field, generator) {
   _.defaults(field, {
     propertyName: field.fieldName,
     path: [field.fieldName],
@@ -364,127 +360,6 @@ export function getEnumValuesWithCustomValues(enumValues) {
       value: matched[2],
     };
   });
-}
-
-export function prepareClientFieldForTemplates(entityWithConfig, field, generator) {
-  if (field.fieldValidateRulesPatternAngular === undefined) {
-    field.fieldValidateRulesPatternAngular = field.fieldValidateRulesPattern
-      ? field.fieldValidateRulesPattern.replace(/"/g, '&#34;')
-      : field.fieldValidateRulesPattern;
-  }
-
-  if (field.fieldValidateRulesPatternReact === undefined) {
-    field.fieldValidateRulesPatternReact = field.fieldValidateRulesPattern
-      ? field.fieldValidateRulesPattern.replace(/'/g, "\\'")
-      : field.fieldValidateRulesPattern;
-  }
-}
-
-export function prepareServerFieldForTemplates(entityWithConfig, field, generator) {
-  if (field.mapstructExpression) {
-    assert.equal(
-      entityWithConfig.dto,
-      MAPSTRUCT,
-      `@MapstructExpression requires an Entity with mapstruct dto [${entityWithConfig.name}.${field.fieldName}].`
-    );
-    // Remove from Entity.java and liquibase.
-    field.transient = true;
-    // Disable update form.
-    field.readonly = true;
-  }
-
-  if (field.id && entityWithConfig.primaryKey) {
-    if (field.autoGenerate === undefined) {
-      field.autoGenerate = !entityWithConfig.primaryKey.composite && [LONG, UUID].includes(field.fieldType);
-    }
-
-    if (!field.autoGenerate) {
-      field.liquibaseAutoIncrement = false;
-      field.jpaGeneratedValue = false;
-      field.autoGenerateByService = false;
-      field.autoGenerateByRepository = false;
-      field.requiresPersistableImplementation = true;
-    } else if (entityWithConfig.databaseType !== SQL) {
-      field.liquibaseAutoIncrement = false;
-      field.jpaGeneratedValue = false;
-      field.autoGenerateByService = field.fieldType === UUID;
-      field.autoGenerateByRepository = !field.autoGenerateByService;
-      field.requiresPersistableImplementation = false;
-      field.readonly = true;
-    } else if (entityWithConfig.reactive) {
-      field.liquibaseAutoIncrement = field.fieldType === LONG;
-      field.jpaGeneratedValue = false;
-      field.autoGenerateByService = !field.liquibaseAutoIncrement;
-      field.autoGenerateByRepository = !field.autoGenerateByService;
-      field.requiresPersistableImplementation = !field.liquibaseAutoIncrement;
-      field.readonly = true;
-    } else {
-      const defaultGenerationType = entityWithConfig.prodDatabaseType === MYSQL ? 'identity' : 'sequence';
-      field.jpaGeneratedValue = field.jpaGeneratedValue || field.fieldType === LONG ? defaultGenerationType : true;
-      field.autoGenerateByService = false;
-      field.autoGenerateByRepository = true;
-      field.requiresPersistableImplementation = false;
-      field.readonly = true;
-      if (field.jpaGeneratedValue === 'identity') {
-        field.liquibaseAutoIncrement = true;
-      }
-    }
-  }
-
-  if (field.fieldNameAsDatabaseColumn === undefined) {
-    const fieldNameUnderscored = _.snakeCase(field.fieldName);
-    const jhiFieldNamePrefix = generator.getColumnName(entityWithConfig.jhiPrefix);
-
-    if (isReservedTableName(fieldNameUnderscored, entityWithConfig.prodDatabaseType)) {
-      if (!jhiFieldNamePrefix) {
-        generator.logger.warn(
-          `The field name '${fieldNameUnderscored}' is regarded as a reserved keyword, but you have defined an empty jhiPrefix. This might lead to a non-working application.`
-        );
-        field.fieldNameAsDatabaseColumn = fieldNameUnderscored;
-      } else {
-        field.fieldNameAsDatabaseColumn = `${jhiFieldNamePrefix}_${fieldNameUnderscored}`;
-      }
-    } else {
-      field.fieldNameAsDatabaseColumn = fieldNameUnderscored;
-    }
-  }
-
-  field.columnName = field.fieldNameAsDatabaseColumn;
-  if (field.unique) {
-    field.uniqueConstraintName = generator.getUXConstraintName(
-      entityWithConfig.entityTableName,
-      field.columnName,
-      entityWithConfig.prodDatabaseType
-    );
-  }
-
-  if (field.fieldInJavaBeanMethod === undefined) {
-    // Handle the specific case when the second letter is capitalized
-    // See http://stackoverflow.com/questions/2948083/naming-convention-for-getters-setters-in-java
-    if (field.fieldName.length > 1) {
-      const firstLetter = field.fieldName.charAt(0);
-      const secondLetter = field.fieldName.charAt(1);
-      if (firstLetter === firstLetter.toLowerCase() && secondLetter === secondLetter.toUpperCase()) {
-        field.fieldInJavaBeanMethod = firstLetter.toLowerCase() + field.fieldName.slice(1);
-      } else {
-        field.fieldInJavaBeanMethod = _.upperFirst(field.fieldName);
-      }
-    } else {
-      field.fieldInJavaBeanMethod = _.upperFirst(field.fieldName);
-    }
-  }
-
-  if (field.fieldValidateRulesPatternJava === undefined) {
-    field.fieldValidateRulesPatternJava = field.fieldValidateRulesPattern
-      ? field.fieldValidateRulesPattern.replace(/\\/g, '\\\\').replace(/"/g, '\\"')
-      : field.fieldValidateRulesPattern;
-  }
-
-  if (field.blobContentTypeText) {
-    field.javaFieldType = 'String';
-  } else {
-    field.javaFieldType = field.fieldType;
-  }
 }
 
 export function fieldToReference(entity, field, pathPrefix = []) {
