+++ generators/bootstrap-application-client/generator.mts
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { preparePostEntityClientDerivedProperties } from '../../utils/entity.mjs';
+import { preparePostEntityClientDerivedProperties } from '../client/support/index.mjs';
 import BaseApplicationGenerator from '../base-application/index.mjs';
 import { GENERATOR_BOOTSTRAP_APPLICATION_BASE } from '../generator-list.mjs';
 import type { ClientApplication } from '../client/types.mjs';
