+++ generators/bootstrap-application-server/generator.mts
@@ -21,12 +21,8 @@ import _ from 'lodash';
 import BaseApplicationGenerator from '../base-application/index.mjs';
 import { GENERATOR_BOOTSTRAP_APPLICATION_BASE } from '../generator-list.mjs';
 import { dockerContainers, javaDependencies } from '../generator-constants.mjs';
-import {
-  loadRequiredConfigIntoEntity,
-  loadRequiredConfigDerivedProperties,
-  prepareEntityServerForTemplates,
-  prepareEntityPrimaryKeyForTemplates,
-} from '../../utils/entity.mjs';
+import { loadRequiredConfigIntoEntity, prepareEntityPrimaryKeyForTemplates } from '../base-application/support/index.mjs';
+import { loadRequiredConfigDerivedProperties, prepareEntity as prepareEntityServerForTemplates } from '../server/support/index.mjs';
 import type { SpringBootApplication } from '../server/types.mjs';
 import { authenticationTypes, fieldTypes } from '../../jdl/jhipster/index.mjs';
 import { prepareField as prepareFieldForLiquibaseTemplates } from '../liquibase/support/index.mjs';
