+++ generators/server/support/index.mts
@@ -20,7 +20,11 @@ export * from './doc.mjs';
 export * from './database.mjs';
 export * from './dependabot-maven.mjs';
 export * from './files.mjs';
-export { getJavaValueGeneratorForType, getPrimaryKeyValue } from './templates/field-values.mjs';
-export { javaBeanCase, buildJavaGet, buildJavaGetter, buildJavaSetter } from './java-formatting.mjs';
+export * from './java-formatting.mjs';
+export { default as prepareEntity } from './prepare-entity.mjs';
+export * from './prepare-entity.mjs';
+export { default as prepareField } from './prepare-field.mjs';
+export * from './string.mjs';
+export * from './templates/field-values.mjs';
 export { default as updateLanguagesTask } from './update-languages.mjs';
 export * from './update-languages.mjs';
